api/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
pydantic
mangum
mangum
